refactor: add ConfigOptions module to store configuration routines
This now exports getValueType so it can be used by other tables.
