+++ src/components/OptionsTable.tsx
@@ -1,11 +1,12 @@
 import {
-  Table, Thead, Tr, Th, Tbody, Td, Link,
+  Table, Thead, Tr, Th, Tbody, Td,
 } from '@chakra-ui/react';
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
 
 import configSchema from '../content/mergify-configuration-openapi.json';
-import { getTypeLink, getTypeDescription } from '../utils/getTypeLink';
+
+import { getValueType, OptionDefinition } from './ConfigOptions';
 
 import InlineCode from './InlineCode';
 import { mdxComponents } from './Page';
@@ -13,13 +14,6 @@ import { mdxComponents } from './Page';
 interface Props {
   /** Name to retrieve its options */
   name: keyof typeof configSchema.definitions;
-}
-
-interface OptionDefinition {
-  valueType: string;
-  description: string;
-  default: string | boolean;
-  $ref: any;
 }
 
 export default function OptionsTable({ name }: Props) {
@@ -41,43 +35,7 @@ export default function OptionsTable({ name }: Props) {
       </Thead>
       <Tbody>
         {Object.entries(options).map(([optionKey, definition]) => {
-          let valueType = null;
-
-          if (definition.type === 'array') {
-            const typeLink = getTypeLink(definition.items.$ref);
-            const typeDescription = (
-              <ReactMarkdown components={mdxComponents as any}>
-                {getTypeDescription(definition.items.$ref)}
-              </ReactMarkdown>
-            );
-
-            if (typeLink !== undefined) {
-              valueType = (
-                <>list of
-                  <Link color="primary" textDecoration="underline" href={typeLink}>
-                    {typeDescription}
-                  </Link>
-                </>
-              );
-            } else valueType = <>list of {typeDescription}</>;
-          } else if (definition.$ref !== undefined) {
-            const typeLink = getTypeLink(definition.$ref);
-            const typeDescription = (
-              <ReactMarkdown components={mdxComponents as any}>
-                {getTypeDescription(definition.$ref)}
-              </ReactMarkdown>
-            );
-
-            if (typeLink !== undefined) {
-              valueType = (
-                <Link color="primary" textDecoration="underline" href={typeLink}>
-                  {typeDescription}
-                </Link>
-              );
-            } else valueType = typeDescription;
-          } else {
-            valueType = definition.type;
-          }
+          const valueType = getValueType(definition);
 
           return (
             <>
