+++ src/components/PullRequestAttributesTable.jsx
@@ -7,7 +7,7 @@ import ReactMarkdown from 'react-markdown';
 
 import configSchema from '../content/mergify-configuration-openapi.json';
 
-import { getTypeLink } from '../utils/getTypeLink';
+import { getTypeLink } from './ConfigOptions';
 
 import InlineCode from './InlineCode';
